Revert to using the old embedding_lookup method
Summary: the new approach is not ready yet as it regresses networks.

Test Plan: all tests pass

Reviewers: #tensorflow, davidn

Reviewed By: #tensorflow, davidn

Differential Revision: https://phabricator.sourcevertex.net/D15581

+++ tensorflow/python/ipu/tests/embedding_lookup_test.py
@@ -122,7 +122,7 @@ class EmbeddingLookupTest(test_util.TensorFlowTestCase):
     with self.session() as sess:
       with ops.device('cpu'):
         x1 = array_ops.placeholder(np.int32, shape=[10])
-        lr = array_ops.placeholder(np.int32, shape=[])
+        lr = array_ops.placeholder(np.float32, shape=[])
 
       def network(x1, lr):
         with variable_scope.variable_scope("vs", use_resource=True):
